looker-to-sheets/requirements.txt
playwright==1.49.1
gspread==6.1.4
google-auth==2.37.0
python-dateutil==2.9.0
